backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0
requests>=2.31.0
